feat: add unit testing for homepage
