--- src/components/header/header.test.js
@@ -1,18 +1,0 @@
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import renderer from 'react-test-renderer';
-import Header from './header';
-
-const MockHeader = () => (
-    <BrowserRouter>
-      <Header />
-    </BrowserRouter>
-  );
-  
-  test('renders correctly', () => {
-    const tree = renderer
-      .create(<MockHeader />)
-      .toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
